requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.6.4
pydantic-core==2.16.3
alembic==1.13.1
httpx==0.27.0
rapidfuzz==3.6.1
